+++ Changelog.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+# [0.0.10] - 2026-01-21
+
+## Changes
+- All local imports now use file extensions for better compatibility with different module systems
+- Moved to vitest for testing framework
+
+# [0.0.9] - 2026-01-04
+
+## Added
+- Added isValidBareDate, isValidBareTime, isValidBareDateTime functions to validate BareDate, BareTime and BareDateTime objects
+
 # [0.0.8] - 2025-10-03
 
 ## Added